Load evapotranspiration and add to time series plots
Now exits with an error on data load if evapotranspiration is not on
the same grid as precipitation.  This way the user can use whatever
method they want to interpolate evapotranspiration onto the
precipitation grid.

Also, add an -e, --plot-evapotranspiration flag to `spowtd plot
time-series` to show an additional panel with evapotranspiration.