Implement stable matching between storms and rises
Work with new datasets demonstrated that sometimes a single storm
could be matched to more than one rise, particularly in noisy water
level time series.  This change set ensures that each storm is matched
to at most 1 rise (and, as before, that each rise is matched to at
most 1 storm).  This condition is enforced by a unique constraint on
storms in the storm-rise matching (in addition to the existing unique
constraint on rises), and is implemented by finding a stable matching
between storms and rises using a Gale-Shapley-type algorithm.  The
user guide has been updated to describe this, and has also been
improved in a few other places.

Regression tests have also been updated due to these changes in the
implementation of storm-rise matching.